requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Django==4.2
django-oauth-toolkit==2.4.0
django-sslserver==0.22
djangorestframework==3.14.0
fernet==1.0.1
gunicorn==22.0.0
h11==0.14.0
html2text==2024.2.26
idna==3.6
jwcrypto==1.5.6
lxml==5.2.1
mysqlclient==2.2.1
oauthlib==3.2.2
packaging==24.1
pyaes==1.6.1
pycparser==2.21
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.2.1
uvicorn==0.30.1
xmltodict==0.13.0
